Move `charts/images.yaml` to `imagevector/images.yaml`
