+++ Dockerfile
@@ -12,7 +12,6 @@ WORKDIR /
 ############# gardener-extension-registry-cache
 FROM base AS registry-cache
 
-COPY charts/images.yaml /charts/
 COPY --from=builder /go/bin/gardener-extension-registry-cache /gardener-extension-registry-cache
 ENTRYPOINT ["/gardener-extension-registry-cache"]
