+++ imagevector/imagevector.go
@@ -15,18 +15,22 @@
 package imagevector
 
 import (
+	_ "embed"
+
 	"github.com/gardener/gardener/pkg/utils/imagevector"
 	"k8s.io/apimachinery/pkg/util/runtime"
-
-	"github.com/gardener/gardener-extension-registry-cache/charts"
 )
 
+// ImagesYAML contains the content of the images.yaml file
+//
+//go:embed images.yaml
+var imagesYAML string
 var imageVector imagevector.ImageVector
 
 func init() {
 	var err error
 
-	imageVector, err = imagevector.Read([]byte(charts.ImagesYAML))
+	imageVector, err = imagevector.Read([]byte(imagesYAML))
 	runtime.Must(err)
 
 	imageVector, err = imagevector.WithEnvOverride(imageVector)
